backend/pyproject.toml
[project]
name = "image-translator-backend"
version = "0.1.0"
description = "FastAPI backend for AI-powered comic, manga, and manhwa image translation."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "SQLAlchemy[asyncio]>=2.0.32",
    "asyncpg>=0.29.0",
    "alembic>=1.13.2",
    "redis>=5.0.8",
    "celery[redis]>=5.4.0",
    "boto3>=1.34.0",
    "Pillow>=10.4.0",
    "opencv-python-headless>=4.10.0.84",
    "numpy>=1.26.4",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.2.0",
    "structlog>=24.4.0",
    "reportlab>=4.2.2",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.24.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.6.3",
    "mypy>=1.11.1",
]
ocr = [
    "easyocr>=1.7.2",
    "paddleocr>=2.8.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
plugins = []
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
